chore: update http/2 tunneling text
Update the text for the HTTP/2 gRPC tunneling warning.
