+++ client/pkg/omnictl/download.go
@@ -61,7 +61,8 @@ func init() {
 	downloadCmd.Flags().StringArrayVar(&downloadCmdFlags.extraKernelArgs, "extra-kernel-args", nil, "Add extra kernel args to the generated installation media")
 	downloadCmd.Flags().StringSliceVar(&downloadCmdFlags.extensions, "extensions", nil, "Generate installation media with extensions pre-installed")
 	downloadCmd.Flags().BoolVar(&downloadCmdFlags.useSiderolinkGRPCTunnel, useSiderolinkGRPCTunnelFlag, false,
-		"Configure Talos to use the SideroLink (WireGuard) gRPC tunnel over HTTP2 for Omni management traffic, instead of UDP. Note that this will add overhead to the traffic.")
+		"Configure Talos to use the SideroLink (WireGuard) gRPC tunnel over HTTP/2 for Omni management traffic, instead of UDP."+
+			" Only enable this if the network blocks UDP packets, as HTTP tunneling adds significant overhead for communications.")
 
 	RootCmd.AddCommand(downloadCmd)
 }
